+++ src/behaviors/hv-select-all/index.ts
@@ -9,7 +9,7 @@ import type {
 } from 'hyperview/src/types';
 import { later, shallowCloneToRoot } from 'hyperview/src/services';
 import { ACTIONS } from 'hyperview/src/types';
-import Logging from 'hyperview/src/core/logging';
+import Logging from 'hyperview/src/services/logging';
 
 export default {
   action: ACTIONS.SELECT_ALL,
