+++ src/components/hv-text-field/index.tsx
@@ -9,7 +9,7 @@ import {
 import type { ElementRef } from 'react';
 import type { KeyboardTypeOptions } from 'react-native';
 import { LOCAL_NAME } from 'hyperview/src/types';
-import Logging from 'hyperview/src/core/logging';
+import Logging from 'hyperview/src/services/logging';
 import { TextInput } from 'react-native';
 import TinyMask from 'hyperview/src/mask';
 import debounce from 'lodash/debounce';
