+++ src/services/render/index.tsx
@@ -6,7 +6,7 @@ import type {
   StyleSheets,
 } from 'hyperview/src/types';
 import { LOCAL_NAME, NODE_TYPE } from 'hyperview/src/types';
-import Logging from 'hyperview/src/core/logging';
+import Logging from 'hyperview/src/services/logging';
 import React from 'react';
 
 export const renderElement = (
